DUNEGPT_polaris_new/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Machine Learning & NLP
sentence-transformers==2.2.2
torch>=2.0.0
torchvision
torchaudio
transformers>=4.21.0
spacy>=3.4.0

# Vector Database
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available

# Document Processing
pdfplumber==0.9.0
python-pptx==0.6.21
beautifulsoup4==4.12.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# CLI
click==8.1.7

# Authentication (OAuth2)
authlib==1.2.1

# Environment Management
python-dotenv==1.0.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.5.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
